[finishes #155529783] Reset password endpoint
